Update README.md
This mini-project focuses on the development of a Tumor Detection System using Machine Learning (ML). The project aims to employ advanced ML techniques to enhance the accuracy and efficiency of tumor detection in medical images. The system utilizes a dataset of medical images to train and test the machine learning model for accurate identification of tumors.